extract state color drawable from squircle drawable
